Installing files should now work for Nextjs
