+++ packages/cli/src/commands/init.ts
@@ -6,19 +6,24 @@ import pathModule from "path";
 import { pathToRegexp } from "path-to-regexp";
 import { simpleGit } from "simple-git";
 import { parse } from "tsconfck";
-import { pathToFileURL } from "url";
+
 import { promptApiKey, promptTriggerUrl } from "../cli/index";
 import { CLOUD_API_URL, CLOUD_TRIGGER_URL, COMMAND_NAME } from "../consts";
-import { TelemetryClient, telemetryClient } from "../telemetry/telemetry";
+import { telemetryClient } from "../telemetry/telemetry";
 import { addDependencies } from "../utils/addDependencies";
 import { pathExists, readJSONFile } from "../utils/fileSystem";
 import { logger } from "../utils/logger";
 import { resolvePath } from "../utils/parseNameAndPath";
-import { renderApiKey } from "../utils/renderApiKey";
 import { renderTitle } from "../utils/renderTitle";
 import { TriggerApi, WhoamiResponse } from "../utils/triggerApi";
 import { frameworkNames, getFramework } from "../frameworks";
 import { getUserPackageManager } from "../utils/getUserPkgManager";
+import {
+  getEnvFilename,
+  setApiKeyEnvironmentVariable,
+  setApiUrlEnvironmentVariable,
+} from "../utils/env";
+import { readPackageJson } from "../utils/readPackageJson";
 
 export type InitCommandOptions = {
   projectPath: string;
@@ -68,11 +73,7 @@ export const initCommand = async (options: InitCommandOptions) => {
   const isTypescriptProject = await detectTypescriptProject(resolvedPath);
   telemetryClient.init.isTypescriptProject(isTypescriptProject, options);
 
-  const optionsAfterPrompts = await resolveOptionsWithPrompts(
-    options,
-    resolvedPath,
-    telemetryClient
-  );
+  const optionsAfterPrompts = await resolveOptionsWithPrompts(options, resolvedPath);
   const apiKey = optionsAfterPrompts.apiKey;
 
   if (!apiKey) {
@@ -102,49 +103,31 @@ export const initCommand = async (options: InitCommandOptions) => {
   const endpointSlug = authorizedKey.project.slug;
   const resolvedOptions: ResolvedOptions = { ...optionsAfterPrompts, endpointSlug };
 
-  const dependencies = await framework.dependencies(resolvedPath, packageManager);
-
-  await addDependencies(resolvedPath, [
-    { name: "@trigger.dev/sdk", tag: "latest" },
-    { name: "@trigger.dev/nextjs", tag: "latest" },
-  ]);
-
+  //install dependencies
+  const dependencies = await framework.dependencies();
+  await addDependencies(resolvedPath, dependencies);
   telemetryClient.init.addedDependencies(resolvedOptions);
 
-  // Setup environment variables
-  await setupEnvironmentVariables(resolvedPath, resolvedOptions);
-
-  const usesSrcDir = await detectUseOfSrcDir(resolvedPath);
-
-  if (usesSrcDir) {
-    logger.info("📁 Detected use of src directory");
+  // Setup environment variables (create a file if there isn't one)
+  let envName = await getEnvFilename(resolvedPath, framework.possibleEnvFilenames());
+  if (!envName) {
+    envName = pathModule.join(resolvedPath, framework.possibleEnvFilenames()[0]!);
+    await fs.writeFile(envName, "");
   }
+  await setApiKeyEnvironmentVariable(resolvedPath, envName, resolvedOptions.apiKey);
+  await setApiUrlEnvironmentVariable(resolvedPath, envName, resolvedOptions.apiUrl);
 
-  const nextJsDir = await detectPagesOrAppDir(resolvedPath, usesSrcDir);
+  const installOptions = {
+    typescript: isTypescriptProject,
+    packageManager,
+    endpointSlug: resolvedOptions.endpointSlug,
+  };
 
-  const routeDir = pathModule.join(resolvedPath, usesSrcDir ? "src" : "");
+  telemetryClient.init.install(resolvedOptions, framework.name, installOptions);
+  await framework.install(resolvedPath, installOptions);
 
-  if (nextJsDir === "pages") {
-    telemetryClient.init.createFiles(resolvedOptions, "pages");
-    await createTriggerPageRoute(
-      resolvedPath,
-      routeDir,
-      resolvedOptions,
-      isTypescriptProject,
-      usesSrcDir
-    );
-  } else {
-    telemetryClient.init.createFiles(resolvedOptions, "app");
-    await createTriggerAppRoute(
-      resolvedPath,
-      routeDir,
-      resolvedOptions,
-      isTypescriptProject,
-      usesSrcDir
-    );
-  }
-
-  await detectMiddlewareUsage(resolvedPath, usesSrcDir);
+  telemetryClient.init.postInstall(resolvedOptions, framework.name, installOptions);
+  await framework.postInstall(resolvedPath, installOptions);
 
   await addConfigurationToPackageJson(resolvedPath, resolvedOptions);
 
@@ -170,15 +153,18 @@ async function printNextSteps(options: ResolvedOptions, authorizedKey: WhoamiRes
 }
 
 async function addConfigurationToPackageJson(path: string, options: ResolvedOptions) {
-  const pkgJsonPath = pathModule.join(path, "package.json");
-  const pkgBuffer = await fs.readFile(pkgJsonPath);
-  const pkgJson = JSON.parse(pkgBuffer.toString());
+  const pkgJson = await readPackageJson(path);
+
+  if (!pkgJson) {
+    throw new Error("Could not find package.json");
+  }
 
   pkgJson["trigger.dev"] = {
     endpointId: options.endpointSlug,
   };
 
   // Write the updated package.json file
+  const pkgJsonPath = pathModule.join(path, "package.json");
   await fs.writeFile(pkgJsonPath, JSON.stringify(pkgJson, null, 2));
 
   logger.success(`✅ Wrote trigger.dev config to package.json`);
@@ -190,8 +176,7 @@ type OptionsAfterPrompts = Required<Omit<InitCommandOptions, "endpointSlug">> & 
 
 const resolveOptionsWithPrompts = async (
   options: InitCommandOptions,
-  path: string,
-  telemetryClient: TelemetryClient
+  path: string
 ): Promise<OptionsAfterPrompts> => {
   const resolvedOptions: InitCommandOptions = { ...options };
 
@@ -279,475 +264,4 @@ async function detectTypescriptProject(path: string): Promise<boolean> {
   } catch (error) {
     return false;
   }
-}
-
-async function detectUseOfSrcDir(path: string): Promise<boolean> {
-  // Detects if the project is using a src directory
-  try {
-    await fs.access(pathModule.join(path, "src"));
-    return true;
-  } catch (error) {
-    return false;
-  }
-}
-
-// Detect the use of pages or app dir in the Next.js project
-// Import the next.config.js file and check for experimental: { appDir: true }
-async function detectPagesOrAppDir(path: string, usesSrcDir = false): Promise<"pages" | "app"> {
-  const nextConfigPath = pathModule.join(path, "next.config.js");
-  const importedConfig = await import(pathToFileURL(nextConfigPath).toString()).catch(() => ({}));
-
-  if (importedConfig?.default?.experimental?.appDir) {
-    return "app";
-  } else {
-    // We need to check if src/app/page.tsx exists
-    // Or app/page.tsx exists
-    // If so then we return app
-    // If not return pages
-
-    const extensionsToCheck = ["jsx", "tsx", "js", "ts"];
-    const basePath = pathModule.join(path, usesSrcDir ? "src" : "", "app", `page.`);
-
-    for (const extension of extensionsToCheck) {
-      const appPagePath = basePath + extension;
-      const appPageExists = await pathExists(appPagePath);
-
-      if (appPageExists) {
-        return "app";
-      }
-    }
-
-    return "pages";
-  }
-}
-
-async function detectMiddlewareUsage(path: string, usesSrcDir = false) {
-  const middlewarePath = pathModule.join(path, usesSrcDir ? "src" : "", "middleware.ts");
-
-  const middlewareExists = await pathExists(middlewarePath);
-
-  if (!middlewareExists) {
-    return;
-  }
-
-  const matcher = await getMiddlewareConfigMatcher(middlewarePath);
-
-  if (!matcher || matcher.length === 0) {
-    logger.warn(
-      `⚠️ ⚠️ ⚠️  It looks like there might be conflicting Next.js middleware in ${pathModule.relative(
-        process.cwd(),
-        middlewarePath
-      )} which can cause issues with Trigger.dev. Please see https://trigger.dev/docs/documentation/guides/platforms/nextjs#middleware`
-    );
-
-    telemetryClient.init.warning("middleware_conflict", { projectPath: path });
-    return;
-  }
-
-  if (matcher.length === 0) {
-    return;
-  }
-
-  if (typeof matcher === "string") {
-    const matcherRegex = pathToRegexp(matcher);
-
-    // Check to see if /api/trigger matches the regex, if it does, then we need to output a warning with a link to the docs to fix it
-    if (matcherRegex.test("/api/trigger")) {
-      logger.warn(
-        `🚨 It looks like there might be conflicting Next.js middleware in ${pathModule.relative(
-          process.cwd(),
-          middlewarePath
-        )} which will cause issues with Trigger.dev. Please see https://trigger.dev/docs/documentation/guides/platforms/nextjs#middleware`
-      );
-      telemetryClient.init.warning("middleware_conflict_api_trigger", { projectPath: path });
-    }
-  } else if (Array.isArray(matcher) && matcher.every((m) => typeof m === "string")) {
-    const matcherRegexes = matcher.map((m) => pathToRegexp(m));
-
-    if (matcherRegexes.some((r) => r.test("/api/trigger"))) {
-      logger.warn(
-        `🚨 It looks like there might be conflicting Next.js middleware in ${pathModule.relative(
-          process.cwd(),
-          middlewarePath
-        )} which will cause issues with Trigger.dev. Please see https://trigger.dev/docs/documentation/guides/platforms/nextjs#middleware`
-      );
-      telemetryClient.init.warning("middleware_conflict", { projectPath: path });
-    }
-  }
-}
-
-async function getMiddlewareConfigMatcher(path: string): Promise<Array<string>> {
-  const fileContent = await fs.readFile(path, "utf-8");
-
-  const regex = /matcher:\s*(\[.*\]|".*")/s;
-  let match = regex.exec(fileContent);
-
-  if (!match) {
-    return [];
-  }
-
-  if (match.length < 2) {
-    return [];
-  }
-
-  let matcherString: string = match[1] as string;
-
-  // Handle array scenario
-  if (matcherString.startsWith("[") && matcherString.endsWith("]")) {
-    matcherString = matcherString.slice(1, -1); // Remove brackets
-    const arrayRegex = /("(?:\\.|[^"\\])*"|'(?:\\.|[^'\\])*')/g;
-    let arrayMatch;
-    const matches: string[] = [];
-    while ((arrayMatch = arrayRegex.exec(matcherString)) !== null) {
-      matches.push((arrayMatch[1] as string).slice(1, -1)); // remove quotes
-    }
-    return matches;
-  } else {
-    // Handle single string scenario
-    return [matcherString.slice(1, -1)]; // remove quotes
-  }
-}
-
-// Find the alias that points to the "src" directory.
-// So for example, the paths object could be:
-// {
-//   "@/*": ["./src/*"]
-// }
-// In this case, we would return "@"
-function getPathAlias(tsconfig: any, usesSrcDir: boolean) {
-  if (!tsconfig.compilerOptions.paths) {
-    return;
-  }
-
-  const paths = tsconfig.compilerOptions.paths;
-
-  const alias = Object.keys(paths).find((key) => {
-    const value = paths[key];
-
-    if (value.length !== 1) {
-      return false;
-    }
-
-    const path = value[0];
-
-    if (usesSrcDir) {
-      return path === "./src/*";
-    } else {
-      return path === "./*";
-    }
-  });
-
-  // Make sure to remove the trailing "/*"
-  if (alias) {
-    return alias.slice(0, -2);
-  }
-
-  return;
-}
-
-async function createTriggerAppRoute(
-  projectPath: string,
-  path: string,
-  options: ResolvedOptions,
-  isTypescriptProject: boolean,
-  usesSrcDir = false
-) {
-  const configFileName = isTypescriptProject ? "tsconfig.json" : "jsconfig.json";
-  const tsConfigPath = pathModule.join(projectPath, configFileName);
-  const { tsconfig } = await parse(tsConfigPath);
-
-  const extension = isTypescriptProject ? ".ts" : ".js";
-  const triggerFileName = `trigger${extension}`;
-  const examplesFileName = `examples${extension}`;
-  const examplesIndexFileName = `index${extension}`;
-  const routeFileName = `route${extension}`;
-
-  const pathAlias = getPathAlias(tsconfig, usesSrcDir);
-  const routePathPrefix = pathAlias ? pathAlias + "/" : "../../../";
-
-  const routeContent = `
-import { createAppRoute } from "@trigger.dev/nextjs";
-import { client } from "${routePathPrefix}trigger";
-
-
-import "${routePathPrefix}jobs";
-
-//this route is used to send and receive data with Trigger.dev
-export const { POST, dynamic } = createAppRoute(client);
-`;
-
-  const triggerContent = `
-import { TriggerClient } from "@trigger.dev/sdk";
-
-export const client = new TriggerClient({
-  id: "${options.endpointSlug}",
-  apiKey: process.env.TRIGGER_API_KEY,
-  apiUrl: process.env.TRIGGER_API_URL,
-});
-  `;
-
-  const jobsPathPrefix = pathAlias ? pathAlias + "/" : "../";
-
-  const jobsContent = `
-import { eventTrigger } from "@trigger.dev/sdk";
-import { client } from "${jobsPathPrefix}trigger";
-
-// Your first job
-// This Job will be triggered by an event, log a joke to the console, and then wait 5 seconds before logging the punchline
-client.defineJob({
-  // This is the unique identifier for your Job, it must be unique across all Jobs in your project
-  id: "example-job",
-  name: "Example Job: a joke with a delay",
-  version: "0.0.1",
-  // This is triggered by an event using eventTrigger. You can also trigger Jobs with webhooks, on schedules, and more: https://trigger.dev/docs/documentation/concepts/triggers/introduction
-  trigger: eventTrigger({
-    name: "example.event",
-  }),
-  run: async (payload, io, ctx) => {
-    // This logs a message to the console
-    await io.logger.info("🧪 Example Job: a joke with a delay");
-    await io.logger.info("How do you comfort a JavaScript bug?");
-    // This waits for 5 seconds, the second parameter is the number of seconds to wait, you can add delays of up to a year
-    await io.wait("Wait 5 seconds for the punchline...", 5);
-    await io.logger.info("You console it! 🤦");
-    await io.logger.info(
-      "✨ Congratulations, You just ran your first successful Trigger.dev Job! ✨"
-    );
-    // To learn how to write much more complex (and probably funnier) Jobs, check out our docs: https://trigger.dev/docs/documentation/guides/create-a-job
-  },
-});
-
-`;
-
-  const examplesIndexContent = `
-// import all your job files here
-
-export * from "./examples"
-`;
-
-  const directories = pathModule.join(path, "app", "api", "trigger");
-  await fs.mkdir(directories, { recursive: true });
-
-  const fileExists = await pathExists(pathModule.join(directories, routeFileName));
-
-  if (fileExists) {
-    logger.info("Skipping creation of app route because it already exists");
-    return;
-  }
-
-  await fs.writeFile(pathModule.join(directories, routeFileName), routeContent);
-
-  logger.success(
-    `✅ Created app route at ${usesSrcDir ? "src/" : ""}app/api/${removeFileExtension(
-      triggerFileName
-    )}/${routeFileName}`
-  );
-
-  const triggerFileExists = await pathExists(pathModule.join(path, triggerFileName));
-
-  if (!triggerFileExists) {
-    await fs.writeFile(pathModule.join(path, triggerFileName), triggerContent);
-
-    logger.success(`✅ Created trigger client at ${usesSrcDir ? "src/" : ""}${triggerFileName}`);
-  }
-
-  const exampleDirectories = pathModule.join(path, "jobs");
-  await fs.mkdir(exampleDirectories, { recursive: true });
-
-  const exampleFileExists = await pathExists(pathModule.join(exampleDirectories, examplesFileName));
-
-  if (!exampleFileExists) {
-    await fs.writeFile(pathModule.join(exampleDirectories, examplesFileName), jobsContent);
-
-    await fs.writeFile(
-      pathModule.join(exampleDirectories, examplesIndexFileName),
-      examplesIndexContent
-    );
-
-    logger.success(`✅ Created example job at ${usesSrcDir ? "src/" : ""}jobs/examples.ts`);
-  }
-}
-
-async function createTriggerPageRoute(
-  projectPath: string,
-  path: string,
-  options: ResolvedOptions,
-  isTypescriptProject: boolean,
-  usesSrcDir = false
-) {
-  const configFileName = isTypescriptProject ? "tsconfig.json" : "jsconfig.json";
-  const tsConfigPath = pathModule.join(projectPath, configFileName);
-  const { tsconfig } = await parse(tsConfigPath);
-
-  const pathAlias = getPathAlias(tsconfig, usesSrcDir);
-  const routePathPrefix = pathAlias ? pathAlias + "/" : "../../";
-
-  const extension = isTypescriptProject ? ".ts" : ".js";
-  const triggerFileName = `trigger${extension}`;
-  const examplesFileName = `examples${extension}`;
-  const examplesIndexFileName = `index${extension}`;
-
-  const routeContent = `
-import { createPagesRoute } from "@trigger.dev/nextjs";
-import { client } from "${routePathPrefix}trigger";
-
-import "${routePathPrefix}jobs";
-
-//this route is used to send and receive data with Trigger.dev
-const { handler, config } = createPagesRoute(client);
-export { config };
-
-export default handler;
-  `;
-
-  const triggerContent = `
-import { TriggerClient } from "@trigger.dev/sdk";
-
-export const client = new TriggerClient({
-  id: "${options.endpointSlug}",
-  apiKey: process.env.TRIGGER_API_KEY,
-  apiUrl: process.env.TRIGGER_API_URL,
-});
-  `;
-
-  const jobsPathPrefix = pathAlias ? pathAlias + "/" : "../";
-
-  const jobsContent = `
-import { eventTrigger } from "@trigger.dev/sdk";
-import { client } from "${jobsPathPrefix}trigger";
-
-// Your first job
-// This Job will be triggered by an event, log a joke to the console, and then wait 5 seconds before logging the punchline
-client.defineJob({
-  // This is the unique identifier for your Job, it must be unique across all Jobs in your project
-  id: "example-job",
-  name: "Example Job: a joke with a delay",
-  version: "0.0.1",
-  // This is triggered by an event using eventTrigger. You can also trigger Jobs with webhooks, on schedules, and more: https://trigger.dev/docs/documentation/concepts/triggers/introduction
-  trigger: eventTrigger({
-    name: "example.event",
-  }),
-  run: async (payload, io, ctx) => {
-    // This logs a message to the console
-    await io.logger.info("🧪 Example Job: a joke with a delay");
-    await io.logger.info("How do you comfort a JavaScript bug?");
-    // This waits for 5 seconds, the second parameter is the number of seconds to wait, you can add delays of up to a year
-    await io.wait("Wait 5 seconds for the punchline...", 5);
-    await io.logger.info("You console it! 🤦");
-    await io.logger.info(
-      "✨ Congratulations, You just ran your first successful Trigger.dev Job! ✨"
-    );
-    // To learn how to write much more complex (and probably funnier) Jobs, check out our docs: https://trigger.dev/docs/documentation/guides/create-a-job
-  },
-});
-`;
-
-  const examplesIndexContent = `
-// import all your job files here
-
-export * from "./examples"
-  `;
-
-  const directories = pathModule.join(path, "pages", "api");
-  await fs.mkdir(directories, { recursive: true });
-
-  // Don't overwrite the file if it already exists
-  const exists = await pathExists(pathModule.join(directories, triggerFileName));
-
-  if (exists) {
-    logger.info("Skipping creation of pages route because it already exists");
-    return;
-  }
-
-  await fs.writeFile(pathModule.join(directories, triggerFileName), routeContent);
-  logger.success(
-    `✅ Created pages route at ${usesSrcDir ? "src/" : ""}pages/api/${triggerFileName}`
-  );
-
-  const triggerFileExists = await pathExists(pathModule.join(path, triggerFileName));
-
-  if (!triggerFileExists) {
-    await fs.writeFile(pathModule.join(path, triggerFileName), triggerContent);
-
-    logger.success(`✅ Created TriggerClient at ${usesSrcDir ? "src/" : ""}${triggerFileName}`);
-  }
-
-  const exampleDirectories = pathModule.join(path, "jobs");
-  await fs.mkdir(exampleDirectories, { recursive: true });
-
-  const exampleFileExists = await pathExists(pathModule.join(exampleDirectories, examplesFileName));
-
-  if (!exampleFileExists) {
-    await fs.writeFile(pathModule.join(exampleDirectories, examplesFileName), jobsContent);
-
-    await fs.writeFile(
-      pathModule.join(exampleDirectories, examplesIndexFileName),
-      examplesIndexContent
-    );
-
-    logger.success(
-      `✅ Created example job at ${usesSrcDir ? "src/" : ""}jobs/examples/${examplesFileName}`
-    );
-  }
-}
-
-async function setupEnvironmentVariables(path: string, options: ResolvedOptions) {
-  if (options.apiKey) {
-    await setupEnvironmentVariable(
-      path,
-      ".env.local",
-      "TRIGGER_API_KEY",
-      options.apiKey,
-      true,
-      renderApiKey
-    );
-  }
-
-  if (options.triggerUrl) {
-    await setupEnvironmentVariable(path, ".env.local", "TRIGGER_API_URL", options.triggerUrl, true);
-  }
-}
-
-async function setupEnvironmentVariable(
-  dir: string,
-  fileName: string,
-  variableName: string,
-  value: string,
-  replaceValue: boolean = true,
-  renderer: (value: string) => string = (value) => value
-) {
-  const path = pathModule.join(dir, fileName);
-  const envFileExists = await pathExists(path);
-
-  if (!envFileExists) {
-    await fs.writeFile(path, "");
-  }
-
-  const envFileContent = await fs.readFile(path, "utf-8");
-
-  if (envFileContent.includes(variableName)) {
-    if (!replaceValue) {
-      logger.info(
-        `☑ Skipping setting ${variableName}=${renderer(value)} because it already exists`
-      );
-      return;
-    }
-    // Update the existing value
-    const updatedEnvFileContent = envFileContent.replace(
-      new RegExp(`${variableName}=.*\\n`, "g"),
-      `${variableName}=${value}\n`
-    );
-
-    await fs.writeFile(path, updatedEnvFileContent);
-
-    logger.success(`✅ Set ${variableName}=${renderer(value)} in ${fileName}`);
-  } else {
-    await fs.appendFile(path, `\n${variableName}=${value}`);
-
-    logger.success(`✅ Added ${variableName}=${renderer(value)} to ${fileName}`);
-  }
-}
-
-function removeFileExtension(filename: string) {
-  return filename.replace(/\.[^.]+$/, "");
 }
