+++ packages/cli/src/utils/removeFileExtension.ts
@@ -1,0 +1,3 @@
+function removeFileExtension(filename: string) {
+  return filename.replace(/\.[^.]+$/, "");
+}
